refactor(server): align station slot hydration and helper boundaries
Keep PATCH station responses consistent with later GET reads by sharing the same redistribution-aware count hydration. Split station query helpers and wiring so derived return-slot math is easier to find and review.
