+++ apps/server/src/domain/stations/index.ts
@@ -5,6 +5,7 @@ export * from "./repository/station-command.repository";
 export * from "./repository/station-query.repository";
 export * from "./services/station-command.live";
 export * from "./services/station-command.service";
+export * from "./services/station-query.live";
 export * from "./services/station-query.service";
 export * from "./services/station-stats.service";
 export type * from "./services/station.service.types";
