+++ apps/server/src/domain/stations/services/station-query.service.ts
@@ -1,23 +1,35 @@
-import { Effect, Layer } from "effect";
+import { Effect, Option } from "effect";
 
-import { StationQueryRepository } from "../repository/station-query.repository";
-import { makeStationQueryService } from "./queries/station.query.service";
+import type { StationQueryRepo } from "../repository/station.repository.types";
+import type { StationQueryService } from "./station.service.types";
 
-export type { StationQueryService } from "./station.service.types";
+import { StationNotFound } from "../errors";
 
-const makeStationQueryServiceEffect = Effect.gen(function* () {
-  const repo = yield* StationQueryRepository;
-  return makeStationQueryService(repo);
-});
+/**
+ * Xây dựng service đọc dữ liệu station cho các query use-case thông thường.
+ *
+ * Service này cố ý không chứa analytics như revenue aggregation; phần đó nằm
+ * trong station stats service riêng.
+ */
+export function makeStationQueryService(repo: StationQueryRepo): StationQueryService {
+  return {
+    listStations: (filter, pageReq) =>
+      repo.listWithOffset(filter, pageReq),
 
-export class StationQueryServiceTag extends Effect.Service<StationQueryServiceTag>()(
-  "StationQueryService",
-  {
-    effect: makeStationQueryServiceEffect,
-  },
-) {}
+    getStationById: id =>
+      Effect.gen(function* () {
+        const maybe = yield* repo.getById(id);
+        if (Option.isNone(maybe)) {
+          return yield* Effect.fail(new StationNotFound({ id }));
+        }
 
-export const StationQueryServiceLive = Layer.effect(
-  StationQueryServiceTag,
-  makeStationQueryServiceEffect.pipe(Effect.map(StationQueryServiceTag.make)),
-);
+        return maybe.value;
+      }),
+
+    listContextExcludingId: excludedId =>
+      repo.listContextExcludingId(excludedId),
+
+    listNearestStations: args =>
+      repo.listNearest(args),
+  };
+}
